+++ sdk/compute/Azure.ResourceManager.Compute/tests/ComputeTestBase.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System.Threading.Tasks;
+using Azure.Core;
 using Azure.Core.TestFramework;
 using Azure.ResourceManager.Resources;
 using Azure.ResourceManager.Resources.Models;
@@ -12,7 +13,7 @@ namespace Azure.ResourceManager.Compute.Tests
 {
     public class ComputeTestBase : ManagementRecordedTestBase<ComputeTestEnvironment>
     {
-        protected Location DefaultLocation => Location.WestUS2;
+        protected AzureLocation DefaultLocation => AzureLocation.WestUS2;
         protected ArmClient Client { get; private set; }
         protected Subscription DefaultSubscription { get; private set; }
         public ComputeTestBase(bool isAsync) : base(isAsync)