+++ sdk/webpubsub/Azure.ResourceManager.WebPubSub/tests/Helpers/WebPubHubServiceClientTestBase.cs
@@ -3,7 +3,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-
+using Azure.Core;
 using Azure.Core.TestFramework;
 using Azure.ResourceManager.Resources;
 using Azure.ResourceManager.Resources.Models;
@@ -62,7 +62,7 @@ namespace Azure.ResourceManager.WebPubSub.Tests.Helpers
             return (await Subscription.GetResourceGroups().CreateOrUpdateAsync(name, new ResourceGroupData(location))).Value;
         }
 
-        protected async Task<WebPubSub> CreateDefaultWebPubSub(string webPubSubName, Location location, ResourceGroup resourceGroup)
+        protected async Task<WebPubSub> CreateDefaultWebPubSub(string webPubSubName, AzureLocation location, ResourceGroup resourceGroup)
         {
             // Create WebPubSub ConfigData
             IList<LiveTraceCategory> categories = new List<LiveTraceCategory>()
@@ -81,7 +81,7 @@ namespace Azure.ResourceManager.WebPubSub.Tests.Helpers
                 new ResourceLogCategory(){ Name = "category1", Enabled = "false" }
             };
 
-            WebPubSubData data = new WebPubSubData(Location.WestUS2)
+            WebPubSubData data = new WebPubSubData(AzureLocation.WestUS2)
             {
                 Sku = new WebPubSubSku("Standard_S1"),
                 LiveTraceConfiguration = new LiveTraceConfiguration("true", categories),