+++ sdk/cosmosdb/Azure.ResourceManager.CosmosDB/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release History
 
-## 1.0.0-beta.2 (2021-12-27)
+## 1.0.0-beta.3 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 1.0.0-beta.2 (2021-12-28)
 
 ### Features Added