+++ sdk/resources/Azure.ResourceManager.Resources/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release History
 
-## 1.0.0-beta.5 (2021-12-27)
+## 1.0.0-beta.6 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 1.0.0-beta.5 (2021-12-28)
 
 ### Features Added